+++ src/stories/inputs/TextField/ValidationPreview.tsx
@@ -8,22 +8,22 @@ const ValidationPreview = () => {
   return (
     <Grid container spacing={4} sx={{ justifyItems: 'flex-start' }}>
       <Grid size={6}>
-        <TextField label="Error" fullWidth error />
+        <TextField label="Error" value='' fullWidth error />
       </Grid>
       <Grid size={6}>
-        <TextField label="Error" helperText="Incorrect input" fullWidth error />
+        <TextField label="Error" value='' helperText="Incorrect input" fullWidth error />
       </Grid>
       <Grid size={6}>
-        <TextField label="Error" variant="outlined" fullWidth error />
+        <TextField label="Error" value='' variant="outlined" fullWidth error />
       </Grid>
       <Grid size={6}>
-        <TextField label="Error" helperText="Incorrect input" variant="outlined" fullWidth error />
+        <TextField label="Error" value='' helperText="Incorrect input" variant="outlined" fullWidth error />
       </Grid>
       <Grid size={6}>
-        <TextField label="Error" variant="filled" fullWidth error />
+        <TextField label="Error" value='' variant="filled" fullWidth error />
       </Grid>
       <Grid size={6}>
-        <TextField label="Error" helperText="Incorrect input" variant="filled" fullWidth error />
+        <TextField label="Error" value='' helperText="Incorrect input" variant="filled" fullWidth error />
       </Grid>
     </Grid>
   )